+++ spring-cloud-sleuth-core/src/main/java/org/springframework/cloud/sleuth/event/ServerReceivedEvent.java
@@ -16,61 +16,19 @@
 
 package org.springframework.cloud.sleuth.event;
 
-import java.util.Objects;
-
 import org.springframework.cloud.sleuth.Span;
-import org.springframework.context.ApplicationEvent;
 
 /**
  * @author Spencer Gibb
  */
 @SuppressWarnings("serial")
-public class ServerReceivedEvent extends ApplicationEvent {
-
-	private final Span parent;
-	private final Span span;
+public class ServerReceivedEvent extends SpanParentContainingEvent {
 
 	public ServerReceivedEvent(Object source, Span span) {
 		this(source, null, span);
 	}
 
 	public ServerReceivedEvent(Object source, Span parent, Span span) {
-		super(source);
-		this.parent = parent;
-		this.span = span;
-	}
-
-	public Span getParent() {
-		return this.parent;
-	}
-
-	public Span getSpan() {
-		return this.span;
-	}
-
-	@Override
-	public boolean equals(Object o) {
-		if (this == o) {
-			return true;
-		}
-		if (o == null || getClass() != o.getClass()) {
-			return false;
-		}
-		ServerReceivedEvent that = (ServerReceivedEvent) o;
-		return Objects.equals(this.parent, that.parent) && Objects
-				.equals(this.span, that.span);
-	}
-
-	@Override
-	public int hashCode() {
-		return Objects.hash(this.parent, this.span);
-	}
-
-	@Override
-	public String toString() {
-		return "ServerReceivedEvent{" +
-				"parent=" + this.parent +
-				", span=" + this.span +
-				'}';
+		super(source, parent, span);
 	}
 }
