+++ spring-cloud-sleuth-core/src/main/java/org/springframework/cloud/sleuth/instrument/TraceKeys.java
@@ -18,7 +18,6 @@ package org.springframework.cloud.sleuth.instrument;
 
 import java.util.Collection;
 import java.util.LinkedHashSet;
-import java.util.Objects;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
@@ -54,9 +53,6 @@ public class TraceKeys {
 
 	private Message message = new Message();
 
-	public TraceKeys() {
-	}
-
 	public Http getHttp() {
 		return this.http;
 	}
@@ -73,35 +69,33 @@ public class TraceKeys {
 		this.message = message;
 	}
 
-	@Override
-	public boolean equals(Object o) {
-		if (this == o) {
-			return true;
-		}
-		if (o == null || getClass() != o.getClass()) {
-			return false;
-		}
-		TraceKeys traceKeys = (TraceKeys) o;
-		return Objects.equals(this.http, traceKeys.http) && Objects
-				.equals(this.message, traceKeys.message);
-	}
-
-	@Override
-	public int hashCode() {
-		return Objects.hash(this.http, this.message);
-	}
-
-	@Override
-	public String toString() {
-		return "TraceKeys{" +
-				"http=" + this.http +
-				", message=" + this.message +
-				'}';
-	}
-
 	public static class Message {
 
 		private Payload payload = new Payload();
+
+		public Payload getPayload() {
+			return this.payload;
+		}
+
+		public String getPrefix() {
+			return this.prefix;
+		}
+
+		public Collection<String> getHeaders() {
+			return this.headers;
+		}
+
+		public void setPayload(Payload payload) {
+			this.payload = payload;
+		}
+
+		public void setPrefix(String prefix) {
+			this.prefix = prefix;
+		}
+
+		public void setHeaders(Collection<String> headers) {
+			this.headers = headers;
+		}
 
 		public static class Payload {
 			/**
@@ -113,42 +107,16 @@ public class TraceKeys {
 			 */
 			private String type = "message/payload-type";
 
-			@Override
-			public boolean equals(Object o) {
-				if (this == o) {
-					return true;
-				}
-				if (o == null || getClass() != o.getClass()) {
-					return false;
-				}
-				Payload payload = (Payload) o;
-				return Objects.equals(this.size, payload.size) && Objects
-						.equals(this.type, payload.type);
-			}
-
-			@Override
-			public int hashCode() {
-				return Objects.hash(this.size, this.type);
-			}
-
-			@Override
-			public String toString() {
-				return "Payload{" +
-						"size='" + this.size + '\'' +
-						", type='" + this.type + '\'' +
-						'}';
-			}
-
 			public String getSize() {
 				return this.size;
 			}
 
-			public void setSize(String size) {
-				this.size = size;
-			}
-
 			public String getType() {
 				return this.type;
+			}
+
+			public void setSize(String size) {
+				this.size = size;
 			}
 
 			public void setType(String type) {
@@ -168,57 +136,6 @@ public class TraceKeys {
 		 */
 		private Collection<String> headers = new LinkedHashSet<String>();
 
-		@Override
-		public boolean equals(Object o) {
-			if (this == o) {
-				return true;
-			}
-			if (o == null || getClass() != o.getClass()) {
-				return false;
-			}
-			Message message = (Message) o;
-			return Objects.equals(this.payload, message.payload) &&
-					Objects.equals(this.prefix, message.prefix) &&
-					Objects.equals(this.headers, message.headers);
-		}
-
-		@Override
-		public int hashCode() {
-			return Objects.hash(this.payload, this.prefix, this.headers);
-		}
-
-		@Override
-		public String toString() {
-			return "Message{" +
-					"payload=" + this.payload +
-					", prefix='" + this.prefix + '\'' +
-					", headers=" + this.headers +
-					'}';
-		}
-
-		public Payload getPayload() {
-			return this.payload;
-		}
-
-		public void setPayload(Payload payload) {
-			this.payload = payload;
-		}
-
-		public String getPrefix() {
-			return this.prefix;
-		}
-
-		public void setPrefix(String prefix) {
-			this.prefix = prefix;
-		}
-
-		public Collection<String> getHeaders() {
-			return this.headers;
-		}
-
-		public void setHeaders(Collection<String> headers) {
-			this.headers = headers;
-		}
 	}
 
 	public static class Http {
@@ -293,115 +210,72 @@ public class TraceKeys {
 		 */
 		private Collection<String> headers = new LinkedHashSet<String>();
 
-		@Override
-		public boolean equals(Object o) {
-			if (this == o) {
-				return true;
-			}
-			if (o == null || getClass() != o.getClass()) {
-				return false;
-			}
-			Http http = (Http) o;
-			return Objects.equals(this.host, http.host) &&
-					Objects.equals(this.method, http.method) &&
-					Objects.equals(this.path, http.path) &&
-					Objects.equals(this.url, http.url) &&
-					Objects.equals(this.statusCode, http.statusCode) &&
-					Objects.equals(this.requestSize, http.requestSize) &&
-					Objects.equals(this.responseSize, http.responseSize) &&
-					Objects.equals(this.prefix, http.prefix) &&
-					Objects.equals(this.headers, http.headers);
-		}
-
-		@Override
-		public int hashCode() {
-			return Objects
-					.hash(this.host, this.method, this.path, this.url, this.statusCode,
-							this.requestSize, this.responseSize, this.prefix,
-							this.headers);
-		}
-
-		@Override
-		public String toString() {
-			return "Http{" +
-					"host='" + this.host + '\'' +
-					", method='" + this.method + '\'' +
-					", path='" + this.path + '\'' +
-					", url='" + this.url + '\'' +
-					", statusCode='" + this.statusCode + '\'' +
-					", requestSize='" + this.requestSize + '\'' +
-					", responseSize='" + this.responseSize + '\'' +
-					", prefix='" + this.prefix + '\'' +
-					", headers=" + this.headers +
-					'}';
-		}
-
 		public String getHost() {
 			return this.host;
-		}
-
-		public void setHost(String host) {
-			this.host = host;
 		}
 
 		public String getMethod() {
 			return this.method;
 		}
 
-		public void setMethod(String method) {
-			this.method = method;
-		}
-
 		public String getPath() {
 			return this.path;
-		}
-
-		public void setPath(String path) {
-			this.path = path;
 		}
 
 		public String getUrl() {
 			return this.url;
 		}
 
-		public void setUrl(String url) {
-			this.url = url;
-		}
-
 		public String getStatusCode() {
 			return this.statusCode;
-		}
-
-		public void setStatusCode(String statusCode) {
-			this.statusCode = statusCode;
 		}
 
 		public String getRequestSize() {
 			return this.requestSize;
 		}
 
-		public void setRequestSize(String requestSize) {
-			this.requestSize = requestSize;
-		}
-
 		public String getResponseSize() {
 			return this.responseSize;
-		}
-
-		public void setResponseSize(String responseSize) {
-			this.responseSize = responseSize;
 		}
 
 		public String getPrefix() {
 			return this.prefix;
 		}
 
-		public void setPrefix(String prefix) {
-			this.prefix = prefix;
-		}
-
 		public Collection<String> getHeaders() {
 			return this.headers;
+		}
+
+		public void setHost(String host) {
+			this.host = host;
+		}
+
+		public void setMethod(String method) {
+			this.method = method;
+		}
+
+		public void setPath(String path) {
+			this.path = path;
+		}
+
+		public void setUrl(String url) {
+			this.url = url;
+		}
+
+		public void setStatusCode(String statusCode) {
+			this.statusCode = statusCode;
+		}
+
+		public void setRequestSize(String requestSize) {
+			this.requestSize = requestSize;
+		}
+
+		public void setResponseSize(String responseSize) {
+			this.responseSize = responseSize;
+		}
+
+		public void setPrefix(String prefix) {
+			this.prefix = prefix;
 		}
 
 		public void setHeaders(Collection<String> headers) {
